Add option to cache on popstate (defaults to true)
When the option is set to false, a popstate request will not be cached.

The original reason behind this is that popstate and Angular compilation of the cached container HTML do not always play well together.
In some cases, Angular throws an error.  Whilst better use of Angular may prevent this, the option added provides a nice quick fix.